requirement.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.4
Faker==5.0.2
future==0.18.2
html5lib==1.1
idna==2.10
isodate==0.6.0
lxml==4.6.2
Pillow==8.0.1
pycparser==2.20
Pyphen==0.10.0
python-dateutil==2.8.1
python-whois==0.7.3
pytz==2020.4
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tinycss2==1.1.0
urllib3==1.26.2
WeasyPrint==52.2
webencodings==0.5.1
zeep==4.0.0
